Bring the site back to working order
The dependencies are very out of date, and its using Angular1 with
a weird project structure.  But the website is now in working order.
Had to override the github-activity component because for some reason
angular wasn't passing the props along.

Moral of the story, site needs to be rewritten in react